requirements.txt
Pillow>=9.0
selenium>=4.0
webdriver-manager>=3.8
pyperclip>=1.8